searching and sorting arrays
A lottery ticket buyer purchases 10 tickets a week, always playing the same 10 five-digit “lucky” numbers.
Write a program that initializes an array or a vector with the lucky numbers and then lets the player enter this week’s winning five-digit number. The program should perform a linear search through the list of the player’s numbers and report whether or not one of the tickets is a winner this week. (See the sample runs below for examples of the expected output when the user enters a winning or a losing number.)